Improve CI/CD: keep build artifacts
This commit extracts the built .deb files from the build image to
keep them as CI/CD artifacts.

The way it works is to put an "intentional" useless entrypoint into
the build-stage of the multistage-build which would "tar" the artifacts
to stdout. The build-stage is used to create an adhoc container which
makes use of that "useless" entrypoint.

Packages are no longer copied. A buildkit bind mount from the
first stage into the final container is used instead.

Change-Id: Ic90acad1dbef49aebf552374d553a32f7d6a0f7b