add UE IP address Pool Identity IE is Assoc Setup and Update Req.
UE IP address Pool Identity was added in Rel.16 to the PFCP Association
Setup Request, but not to the Setup Response or Update Request.

It seems that this was an editorial overside and the IE was supposed
to be present in those PFCP messages. In any case, it makes sense to
have them there.